requirements.txt
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
Flask==3.0.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
protobuf==4.21.12
pycparser==2.21
requests==2.31.0
urllib3==2.1.0
Werkzeug==3.0.1
zipp==3.17.0
